extracts test helpers away from shipping code
