+++ packages/openapi-2-kong/tsconfig.build.json
@@ -11,6 +11,7 @@
     "src"
   ],
   "exclude": [
-    "**/*.test.ts"
+    "**/*.test.ts",
+    "jest"
   ]
 }
